experiments/new-approach/pyproject.toml
[project]
name = "new-approach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.44",
    "pandas>=2.2.3",
    "sentence-transformers>=4.0.1",
]
